feat(all): split module-loader into an amd and a commonjs package
