+++ packages/demos/package.json
@@ -7,7 +7,8 @@
     "@blueprintjs/core": "^3.10.0",
     "@feature-hub/core": "^0.10.0",
     "@feature-hub/history-service": "^0.10.0",
-    "@feature-hub/module-loader": "^0.10.0",
+    "@feature-hub/module-loader-amd": "0.0.0",
+    "@feature-hub/module-loader-commonjs": "0.0.0",
     "@feature-hub/react": "^0.10.0",
     "@feature-hub/server-renderer": "^0.10.0",
     "@types/express": "^4.16.0",
