+++ packages/demos/src/module-loader-amd/feature-app.tsx
