+++ packages/demos/src/module-loader-amd/index.test.ts
