+++ packages/demos/src/module-loader-amd/integrator.tsx
@@ -1,5 +1,5 @@
 import {FeatureAppManager, FeatureServiceRegistry} from '@feature-hub/core';
-import {defineExternals, loadAmdModule} from '@feature-hub/module-loader';
+import {defineExternals, loadAmdModule} from '@feature-hub/module-loader-amd';
 import {FeatureAppLoader} from '@feature-hub/react';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
