+++ packages/demos/src/module-loader-amd/webpack-config.ts
