+++ packages/demos/src/module-loader-commonjs/feature-app.tsx
