+++ packages/demos/src/module-loader-commonjs/integrator.node.tsx
@@ -1,5 +1,5 @@
 import {FeatureAppManager, FeatureServiceRegistry} from '@feature-hub/core';
-import {loadCommonJsModule} from '@feature-hub/module-loader/lib/cjs/node';
+import {loadCommonJsModule} from '@feature-hub/module-loader-commonjs';
 import {FeatureAppLoader} from '@feature-hub/react';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom/server';
