+++ packages/demos/src/module-loader-commonjs/integrator.ts
