+++ packages/demos/src/module-loader-commonjs/webpack-config.ts
