+++ packages/module-loader-amd/README.md
@@ -1,22 +1,22 @@
-# @feature-hub/module-loader
+# @feature-hub/module-loader-amd
 
 [![Package Version][package-badge]][package-npm]
 [![Website][website-badge]][website] [![API][api-badge]][api]
 
-A universal module loader.
+A AMD module loader.
 
 ## Installation
 
 ### Using Yarn
 
 ```sh
-yarn add @feature-hub/module-loader
+yarn add @feature-hub/module-loader-amd
 ```
 
 ### Using NPM
 
 ```sh
-npm install @feature-hub/module-loader
+npm install @feature-hub/module-loader-amd
 ```
 
 ---
@@ -24,12 +24,12 @@ npm install @feature-hub/module-loader
 Copyright (c) 2018-2019 SinnerSchrader Deutschland GmbH. Released under the
 terms of the [MIT License][license].
 
-[api]: https://feature-hub.io/@feature-hub/module-loader/
+[api]: https://feature-hub.io/@feature-hub/module-loader-amd/
 [api-badge]:
-  https://img.shields.io/badge/API-%40feature--hub%2Fmodule--loader-%23ea3458.svg
+  https://img.shields.io/badge/API-%40feature--hub%2Fmodule--loader--amd-%23ea3458.svg
 [license]: https://github.com/sinnerschrader/feature-hub/blob/master/LICENSE
-[package-badge]: https://img.shields.io/npm/v/@feature-hub/module-loader.svg
-[package-npm]: https://www.npmjs.com/package/@feature-hub/module-loader
+[package-badge]: https://img.shields.io/npm/v/@feature-hub/module-loader-amd.svg
+[package-npm]: https://www.npmjs.com/package/@feature-hub/module-loader-amd
 [website]: https://feature-hub.io/
 [website-badge]:
   https://img.shields.io/badge/Website-feature--hub.io-%23500dc5.svg
