+++ packages/module-loader-amd/src/index.ts
@@ -5,11 +5,6 @@ export interface Externals {
   readonly [moduleName: string]: unknown;
 }
 
-/**
- * ```js
- * import {defineExternals} from '@feature-hub/module-loader';
- * ```
- */
 export function defineExternals(externals: Externals): void {
   for (const moduleName of Object.keys(externals)) {
     const external = externals[moduleName];
@@ -18,11 +13,6 @@ export function defineExternals(externals: Externals): void {
   }
 }
 
-/**
- * ```js
- * import {loadAmdModule} from '@feature-hub/module-loader';
- * ```
- */
 export const loadAmdModule: ModuleLoader = async (
   url: string
 ): Promise<unknown> => SystemJS.import(url);
