+++ packages/module-loader-amd/tsconfig.json
