+++ packages/module-loader-amd/tslint.json
