+++ packages/module-loader-commonjs/.size-limit.js
@@ -1,0 +1,8 @@
+// @ts-check
+
+module.exports = [
+  {
+    path: 'lib/index.js',
+    limit: '8 KB'
+  }
+];
