+++ packages/module-loader-commonjs/README.md
@@ -1,0 +1,36 @@
+# @feature-hub/module-loader-commonjs
+
+[![Package Version][package-badge]][package-npm]
+[![Website][website-badge]][website] [![API][api-badge]][api]
+
+A CommonJS module loader.
+
+## Installation
+
+### Using Yarn
+
+```sh
+yarn add @feature-hub/module-loader-commonjs
+```
+
+### Using NPM
+
+```sh
+npm install @feature-hub/module-loader-commonjs
+```
+
+---
+
+Copyright (c) 2018-2019 SinnerSchrader Deutschland GmbH. Released under the
+terms of the [MIT License][license].
+
+[api]: https://feature-hub.io/@feature-hub/module-loader-commonjs/
+[api-badge]:
+  https://img.shields.io/badge/API-%40feature--hub%2Fmodule--loader--commonjs-%23ea3458.svg
+[license]: https://github.com/sinnerschrader/feature-hub/blob/master/LICENSE
+[package-badge]:
+  https://img.shields.io/npm/v/@feature-hub/module-loader-commonjs.svg
+[package-npm]: https://www.npmjs.com/package/@feature-hub/module-loader-commonjs
+[website]: https://feature-hub.io/
+[website-badge]:
+  https://img.shields.io/badge/Website-Feature%20Hub-%234502da.svg
