+++ packages/module-loader-commonjs/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@feature-hub/module-loader-commonjs",
+  "version": "0.0.0",
+  "description": "A FeatureAppManager-compatible CommonJS module loader.",
+  "homepage": "https://feature-hub.io/",
+  "bugs": {
+    "url": "https://github.com/sinnerschrader/feature-hub/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/sinnerschrader/feature-hub.git"
+  },
+  "license": "MIT",
+  "author": "SinnerSchrader Deutschland GmbH",
+  "files": [
+    "lib",
+    "!__tests__"
+  ],
+  "main": "lib/cjs/index.js",
+  "module": "lib/esm/index.js",
+  "typings": "lib/cjs/index.d.ts",
+  "dependencies": {
+    "@feature-hub/core": "^0.10.0",
+    "fs-extra": "^7.0.1",
+    "node-fetch": "^2.3.0"
+  },
+  "devDependencies": {
+    "@types/fs-extra": "^5.0.4",
+    "@types/node-fetch": "^2.1.4"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
