+++ packages/module-loader-commonjs/tsconfig.cjs.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "paths": {},
+    "outDir": "lib/cjs",
+    "rootDir": "src"
+  }
+}
