+++ packages/module-loader-commonjs/tsconfig.esm.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "es6",
+    "paths": {},
+    "outDir": "lib/esm",
+    "rootDir": "src"
+  }
+}
