--- packages/module-loader/CHANGELOG.md
@@ -1,107 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [0.10.0](https://github.com/sinnerschrader/feature-hub/compare/v0.9.0...v0.10.0) (2018-12-20)
-
-
-### Features
-
-* **module-loader:** use standard package entry for node module loader ([#203](https://github.com/sinnerschrader/feature-hub/issues/203)) ([e380857](https://github.com/sinnerschrader/feature-hub/commit/e380857))
-
-
-
-
-
-# [0.9.0](https://github.com/sinnerschrader/feature-hub/compare/v0.8.0...v0.9.0) (2018-12-20)
-
-
-### Features
-
-* **history-service:** remove history package type exports ([#197](https://github.com/sinnerschrader/feature-hub/issues/197)) ([1694769](https://github.com/sinnerschrader/feature-hub/commit/1694769))
-
-
-
-
-
-# [0.8.0](https://github.com/sinnerschrader/feature-hub/compare/v0.7.0...v0.8.0) (2018-12-19)
-
-**Note:** Version bump only for package @feature-hub/module-loader
-
-
-
-
-
-# [0.7.0](https://github.com/sinnerschrader/feature-hub/compare/v0.6.0...v0.7.0) (2018-12-14)
-
-**Note:** Version bump only for package @feature-hub/module-loader
-
-
-
-
-
-# [0.6.0](https://github.com/sinnerschrader/feature-hub/compare/v0.5.1...v0.6.0) (2018-12-13)
-
-**Note:** Version bump only for package @feature-hub/module-loader
-
-
-
-
-
-# [0.5.0](https://github.com/sinnerschrader/feature-hub/compare/v0.4.0...v0.5.0) (2018-12-11)
-
-**Note:** Version bump only for package @feature-hub/module-loader
-
-
-
-
-
-# [0.4.0](https://github.com/sinnerschrader/feature-hub/compare/v0.3.0...v0.4.0) (2018-12-03)
-
-**Note:** Version bump only for package @feature-hub/module-loader
-
-
-
-
-
-# [0.3.0](https://github.com/sinnerschrader/feature-hub/compare/v0.2.0...v0.3.0) (2018-11-30)
-
-
-### Bug Fixes
-
-* **module-loader:** move [@feature-hub](https://github.com/feature-hub)/core to dependencies ([9d1b0f7](https://github.com/sinnerschrader/feature-hub/commit/9d1b0f7))
-
-
-### Features
-
-* **examples:** setup examples workspace ([#81](https://github.com/sinnerschrader/feature-hub/issues/81)) ([170bfb4](https://github.com/sinnerschrader/feature-hub/commit/170bfb4))
-
-
-
-
-
-# [0.2.0](https://github.com/sinnerschrader/feature-hub/compare/v0.1.1...v0.2.0) (2018-11-28)
-
-**Note:** Version bump only for package @feature-hub/module-loader
-
-
-
-
-
-## [0.1.1](https://github.com/sinnerschrader/feature-hub/compare/v0.1.0...v0.1.1) (2018-11-28)
-
-**Note:** Version bump only for package @feature-hub/module-loader
-
-
-
-
-
-# 0.1.0 (2018-11-28)
-
-
-### Features
-
-* **all:** merge all React packages, closes [#30](https://github.com/sinnerschrader/feature-hub/issues/30) ([#31](https://github.com/sinnerschrader/feature-hub/issues/31)) ([0f6fdd6](https://github.com/sinnerschrader/feature-hub/commit/0f6fdd6))
-* **all:** merge module loaders to one universal package ([#36](https://github.com/sinnerschrader/feature-hub/issues/36)) ([e583b9a](https://github.com/sinnerschrader/feature-hub/commit/e583b9a))
-* **module-loader:** rename module loader functions, closes [#49](https://github.com/sinnerschrader/feature-hub/issues/49) ([#51](https://github.com/sinnerschrader/feature-hub/issues/51)) ([f9be3cd](https://github.com/sinnerschrader/feature-hub/commit/f9be3cd))
